+++ pychron/pyscripts/valve_pyscript.py
@@ -5,7 +5,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#   http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -71,7 +71,7 @@ class ValvePyScript(PyScript):
     @verbose_skip
     @named_register('open')
     def _m_open(self, name=None, description=''):
-
+        st = time.time()
         if description is None:
             description = '---'
 
@@ -80,8 +80,9 @@ class ValvePyScript(PyScript):
         result = self._manager_action([('open_valve', (name,), dict(
             mode='script',
             description=description))], protocol=ELPROTOCOL)
-
-        self.debug('---------------------------------------- open {} ({}) result={}'.format(name, description, result))
+        et = time.time() - st
+        self.debug('---------------------------------------- open {} ({}) result={}, '
+                   'time={:0.2f} sec'.format(name, description, result, et))
         if result is not None:
             self._finish_valve_change('open', result, name, description)